fix(core): make inline event handlers compatible with Trusted Types
Inline event handlers are allowed in Angular templates as long as their
values are constant. However, specifying such attributes currently leads
to Trusted Types violations being generated, as Angular passes their
constant values as plain strings to setAttribute, which in this case is
a Trusted Types sink. This is made worse by a Chromium bug
(https://crbug.com/993268) which treats any attribute starting with `on`
as a Trusted Types sink, causing a Trusted Types violation when any
custom attribute starting with `on` is specified in an Angular template.

Fix this by promoting constant (and thus safe) values of inline event
handlers to TrustedScript before they are passed to the setAttribute
call. Note that all attributes that start with 'on' (e.g. 'onload' or
'onerror') are promoted to a TrustedScript, even though they are not
necessarily an inline event handler (e.g. 'online' or 'one'). This is
not security sensitive and will essentially be a no-op as the values get
stringified. This is done to work around the Chromium bug and minimize
code size impact, which would otherwise require listing all valid inline
event handler names in Angular's DOM schema.

Developers interested in restricting usage of inline event handlers in
their templates will still be able to do so by setting a Content
Security Policy that does not include the 'unsafe-inline' directive.